+++ lib/babe/median_test.go
@@ -1,0 +1,62 @@
+package babe
+
+import (
+	"testing"
+)
+
+func TestMedian_OddLength(t *testing.T) {
+	us := []uint64{3, 2, 1, 4, 5}
+	res, err := median(us)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var expected uint64 = 3
+
+	if res != expected {
+		t.Errorf("Fail: got %v expected %v\n", res, expected)
+	}
+
+}
+
+func TestMedian_EvenLength(t *testing.T) {
+	us := []uint64{1, 4, 2, 4, 5, 6}
+	res, err := median(us)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var expected uint64 = 4
+
+	if res != expected {
+		t.Errorf("Fail: got %v expected %v\n", res, expected)
+	}
+
+}
+
+func TestSlotOffset_Failing(t *testing.T) {
+	var st uint64 = 1000001
+	var se uint64 = 1000000
+
+	_, err := slotOffset(st, se)
+	if err == nil {
+		t.Fatal("Fail: did not err for c>1")
+	}
+
+}
+
+func TestSlotOffset(t *testing.T) {
+	var st uint64 = 1000000
+	var se uint64 = 1000001
+
+	res, err := slotOffset(st, se)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	var expected uint64 = 1
+
+	if res != expected {
+		t.Errorf("Fail: got %v expected %v\n", res, expected)
+	}
+}
